requirements.txt
bcrypt==3.2.0
Django==4.0.3
django-cors-headers==3.11.0
mysqlclient==2.1.0
PyJWT==2.3.0
PyMySQL==1.0.2
